# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Church Records Backend REST API
host: localhost:10120
...
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /admin:
    # binds a127 app logic to a route
    x-swagger-router-controller: admin.controller
    get:
      tags: [Admin]
      description: Returns list of admin from Admin Collection
      # used as the method name of the controller
      operationId: getAllAdmin
      responses:
        "200":
          description: Admin Found
          schema:
            type: array
            items:
              $ref: "#/definitions/admin"
        "404":
          description: Empty Collection
          schema:
            $ref: "#/definitions/message"
        # responses may fall through to errors
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
    post:
      tags: [Admin]
      description: Creates new admin entry
      consumes:
        - application/json
      produces:
        - application/json
      operationId: createAdmin
      parameters:
        - name: admindata
          required: true
          in: body
          description: a new admin details
          schema:
            $ref: "#/definitions/adminreg"
      responses:
        "200":
          description: New Admin Created
          schema:
            $ref: "#/definitions/admin"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
    delete:
      tags: [Admin]
      description: Remove all admins in database
      operationId: removeAllAdmin
      responses:
        "200":
          description: Successfully removed all admins
          schema:
            $ref: "#/definitions/message"
        default:
          description: Error
          schema:
            $ref: "#/definitions/message"
  /admin/{adminid}:
    x-swagger-router-controller: admin.controller
    get:
      tags: [Admin]
      description: Return single admin by given id
      operationId: getSingleAdmin
      parameters:
        - name: adminid
          required: true
          in: path
          type: string
      responses:
        "200":
          description: Admin Found
          schema:
            $ref: "#/definitions/admin"
        "404":
          description: Admin not Found
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
    delete:
      tags: [Admin]
      description: Remove single admin by given id
      operationId: removeSingleAdmin
      parameters:
        - name: adminid
          required: true
          in: path
          type: string
      responses:
        "200":
          description: Successfully removed
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
    patch:
      tags: [Admin]
      description: Update by admin id
      operationId: updateSingleAdmin
      parameters:
        - name: adminid
          in: path
          required: true
          type: string
        - name: admin
          in: body
          description: The Updated admin info
          required: true
          schema:
            $ref: "#/definitions/admin"
      responses:
        "200":
          description: Admin Found and Updated
          schema:
            $ref: "#/definitions/admin"
        "404":
          description: Admin not Found
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
# Authentication Marchavel goes here
  /authentication:
    x-swagger-router-controller: authentication.controller
    post:
      tags: [Authentication]
      description: authenticate users
      operationId: authenticateUser
      parameters:
        - name: authenticatedata
          required: true
          in: body
          schema:
            $ref: "#/definitions/login"
      responses:
        "200":
          description: believer Found
          schema:
            $ref: "#/definitions/authres"
        "404":
          description: believer Not Found
          schema:
            $ref: "#/definitions/message"

        "500":
          description: believer authenticated
          schema:
            $ref: "#/definitions/message"
# Authentication Marchavel goes here
  /authentication/checkusercode:
    x-swagger-router-controller: authentication.controller
    post:
      tags: [Authentication]
      description: Check if believer exist
      operationId: checkCode
      parameters:
        - name: usercode
          required: true
          in: body
          schema:
            $ref: "#/definitions/usercode"
      responses:
        "200":
          description: believer Found
          schema:
            $ref: "#/definitions/message"

        "404":
          description: believer Not Found
          schema:
            $ref: "#/definitions/message"

        "500":
          description: believer authenticated
          schema:
            $ref: "#/definitions/message"
# Path Defition to update Password
  /authentication/updatepassword:
    x-swagger-router-controller: authentication.controller
    post:
      tags: [Authentication]
      description: Updates Password
      operationId: updatePassword
      parameters:
        - name: profiledata
          in: body
          description: The Updated info
          required: true
          schema:
            $ref: "#/definitions/logindata"
      responses:
        "200":
          description: User Found and Updated
          schema:
            $ref: "#/definitions/message"
        "404":
          description: User not Found
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
# The End of Authentication Marchavel

# Path Defition to update ProfilePic
...
  /agegroup:
    # binds a127 app logic to a route
    x-swagger-router-controller: agegroup.controller
    get:
      tags: [Agegroup]
      description: Returns list of agegroups from Agegroup Collection
      # used as the method name of the controller
      operationId: getAllAgegroup
      responses:
        "200":
          description: Agegroup Found
          schema:
            type: array
            items:
              $ref: "#/definitions/agegroup"
        "404":
          description: Empty Collection
          schema:
            $ref: "#/definitions/message"
        # responses may fall through to errors
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
    post:
      tags: [Agegroup]
      description: Creates new agegroup entry
      consumes:
        - application/json
      produces:
        - application/json
      operationId: createAgegroup
      parameters:
        - name: agegroupdata
          required: true
          in: body
          description: a new agegroup details
          schema:
            $ref: "#/definitions/agegroup"
      responses:
        "200":
          description: New Odd Created
          schema:
            $ref: "#/definitions/agegroup"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
    delete:
      tags: [Agegroup]
      description: Remove all agegroups in database
      operationId: removeAllAgegroup
      responses:
        "200":
          description: Successfully removed all agegroups
          schema:
            $ref: "#/definitions/message"
        default:
          description: Error
          schema:
            $ref: "#/definitions/message"
  /agegroup/{agegroupid}:
    x-swagger-router-controller: agegroup.controller
    get:
      tags: [Agegroup]
      description: Return single agegroup by given id
      operationId: getSingleAgegroup
      parameters:
        - name: agegroupid
          required: true
          in: path
          type: string
      responses:
        "200":
          description: Agegroup Found
          schema:
            $ref: "#/definitions/agegroup"
        "404":
          description: Agegroup not Found
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
    delete:
      tags: [Agegroup]
      description: Remove single agegroup by given id
      operationId: removeSingleAgegroup
      parameters:
        - name: agegroupid
          required: true
          in: path
          type: string
      responses:
        "200":
          description: Successfully removed
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
    patch:
      tags: [Agegroup]
      description: Update by agegroup id
      operationId: updateSingleAgegroup
      parameters:
        - name: agegroupid
          in: path
          required: true
          type: string
        - name: agegroup
          in: body
          description: The Updated agegroup info
          required: true
          schema:
            $ref: "#/definitions/agegroup"
      responses:
        "200":
          description: Agegroup Found and Updated
          schema:
            $ref: "#/definitions/agegroup"
        "404":
          description: Agegroup not Found
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
# The End of Agegroups
...
definitions:
...
  authres:
    properties:
      _id:
        type: string
      believercode:
        type: string
      admincode:
        type: string
      firstname:
        type: string
      middlename:
        type: string
      lastname:
        type: string
      phonenumber:
        type: string
      device_id:
        type: string
      diocode:
        type: string
      dioname:
        type: string
      churchcode:
        type: string
      churchname:
        type: string
      communitycode:
        type: string
      communityname:
        type: string
      profilepicture:
        type: string
      email:
        type: string
      gender:
        type: string
      dob:
        type: object
        properties:
          year:
            type: number
          month:
            type: number
          day:
            type: number
      status:
        type: string
      healthinfo:
        type: string
      nationality:
        type: string
      jobtitle:
        type: string
      regdate:
        type: string
        format: datetime
      baptismstatus:
        type: string
      maritalstatus:
        type: string
      participationlevel:
        type: array
        items:
          type: string
      residenceadress:
        type: object
        properties:
          street:
            type: string
          road:
            type: string
          district:
            type: string
          region:
            type: string
      relatives:
        type: array
        items:
          type: object
          properties:
            relation:
              type: string
            person:
              type: string
      hobbies:
        type: array
        items:
          type: string
      created_at:
        type: string
        format: datetime
      updated_at:
        type: string
        format: datetime
# Believer Model Definition
...
  adminreg:
    required:
      - admincode
      - title
      - status
      - password
      - firstname
      - lastname
      - gender
      - dob
    properties:
      _id:
        type: string
      admincode:
        type: string
      password:
        type: string
      title:
        type: string
      firstname:
        type: string
      middlename:
        type: string
      lastname:
        type: string
      phonenumber:
        type: string
      device_id:
        type: string
      diocode:
        type: string
      dioname:
        type: string
      churchid:
        type: string
      churchcode:
        type: string
      churchname:
        type: string
      communitycode:
        type: string
      communityname:
        type: string
      profilepicture:
        type: string
      email:
        type: string
      gender:
        type: string
      dob:
        type: object
        properties:
          year:
            type: number
          month:
            type: number
          day:
            type: number
      status:
        type: string
      healthinfo:
        type: string
      nationality:
        type: string
      jobtitle:
        type: string
      regdate:
        type: string
        format: datetime
      baptismstatus:
        type: string
      maritalstatus:
        type: string
      participationlevel:
        type: array
        items:
          type: string
      residenceadress:
        type: object
        properties:
          street:
            type: string
          road:
            type: string
          district:
            type: string
          region:
            type: string
      relatives:
        type: array
        items:
          type: object
          properties:
            relation:
              type: string
            person:
              type: string
      hobbies:
        type: array
        items:
          type: string
      created_at:
        type: string
        format: datetime
      updated_at:
        type: string
        format: datetime
# Admin Model Definition
  admin:
    properties:
      _id:
        type: string
      admincode:
        type: string
      title:
        type: string
      firstname:
        type: string
      middlename:
        type: string
      lastname:
        type: string
      phonenumber:
        type: string
      device_id:
        type: string
      diocode:
        type: string
      dioname:
        type: string
      churchid:
        type: string
      churchcode:
        type: string
      churchname:
        type: string
      communitycode:
        type: string
      communityname:
        type: string
      profilepicture:
        type: string
      email:
        type: string
      gender:
        type: string
      dob:
        type: object
        properties:
          year:
            type: number
          month:
            type: number
          day:
            type: number
      status:
        type: string
      healthinfo:
        type: string
      nationality:
        type: string
      jobtitle:
        type: string
      regdate:
        type: string
        format: datetime
      baptismstatus:
        type: string
      maritalstatus:
        type: string
      participationlevel:
        type: array
        items:
          type: string
      residenceadress:
        type: object
        properties:
          street:
            type: string
          road:
            type: string
          district:
            type: string
          region:
            type: string
      relatives:
        type: array
        items:
          type: object
          properties:
            relation:
              type: string
            person:
              type: string
      hobbies:
        type: array
        items:
          type: string
      created_at:
        type: string
        format: datetime
      updated_at:
        type: string
        format: datetime
#Login Model Defition
  login:
    properties:
      username:
        type: string
      password:
        type: string
#Update Password Model Defition
  logindata:
    required:
      - username
      - password
      - newpassword
    properties:
      username:
        type: string
      password:
        type: string
      newpassword:
        type: string
# Believer Existance Data
  usercode:
    properties:
      username:
        type: string

# General Response Message
  message:
    required:
      - success
      - message
    properties:
      success:
        type: boolean
      message:
        type: string
# Info Model Definition
...
  agegroup:
    required:
      - agegroupname
      - agefrom
      - ageto
      - description
    properties:
      _id:
        type: string
      agegroupname:
        type: string
      agefrom:
        type: number
      ageto:
        type: number
      description:
        type: string
      created_at:
        type: string
        format: datetime
      updated_at:
        type: string
        format: datetime